Added :delay_channels list to configs
delay_channels lets you set a list of channels that should
be delayed.  If this is given, then only channels that
match the option will be stored in Redis.  Other channels
will be sent through Faye immediately without queueing
in Redis waiting for a client.

This is useful if you are using the same Faye instance
for both delayed messages and a general Faye message
queue.